feat(web): consolidate Library, Evals+Compare, Connections via tabs
PR-4 of the unified-IA spec. The new workspace homes from PR-3 stop being redirect placeholders and start hosting the consolidated surfaces under one click each.

frontend/app/dashboard/library/page.tsx — Prompts + Knowledge as tabs.
  Imports PromptsPage and KnowledgePage as-is from their existing routes and
  hosts them under one Library workspace home. Tab persistence via
  ?tab=prompts|knowledge so refreshes / bookmarks land on the right pane.
  Old /dashboard/prompts and /dashboard/knowledge routes stay live as
  deep-link targets (sidebar's "Library" lights for both per PR-3's
  active-state rules).

frontend/app/dashboard/connections/page.tsx — Providers / MCP / Targets /
  Computer-Use config as tabs. Same composition pattern as Library, hosting
  the existing standalone pages inline. The Computer-Use *live view* deep
  link still lives at /dashboard/computer-use; this tab is the config
  surface for it (honours the spec's "don't bury the flagship" guard).

frontend/app/dashboard/evals/page.tsx — Suites + Compare as tabs.
  The previous evals page body extracted into ./SuitesView.tsx (named
  export, identical behaviour); the new page.tsx renders Tabs with
  SuitesView and ComparePage. /dashboard/compare keeps working as a
  deep-link destination.

Vocabulary alignment confirmed (matches PR-2 CLI):
  - Providers (not Models)
  - Team (not Teams) — already settled in PR-3 sidebar
  - Library (Prompts + Knowledge)
  - Connections (Providers + MCP + Targets + Computer-Use)

Tab state lives in URL search params (?tab=...) — same pattern across all
three pages, so back/forward, bookmarks, and shared links work naturally.

What's deferred:
  - Spec §4 also asks for the Studio agent-detail enhancement (surface the
    blueprint and attached prompts/knowledge inline). That's a bigger touch
    in the agent detail view and warrants its own PR. Tracked as a
    follow-up; the workspace consolidation here is the higher-leverage
    win the spec calls out ("cheap, high-impact merges").

Acceptance (spec §PR-4):
  - ✅ Library / Evals / Connections — each merged surface reachable in
    ≤1 click from its workspace
  - ✅ Old standalone pages still resolve (sidebar deep links + bookmarks)
  - ✅ Web labels match CLI namespace names (Providers, Team, Library,
    Connections)

Frontend: 34/34 tests green, lint clean, type-check clean.
